Fixes doctest misconfiguration
fixes reconfiguration introduced in #73 that caused doctests for modules not to execute.

An unfortunate side-effect of specifying the starting directory for test discovery, pointing it at a directory outside our Module space, is that doctests are not discovered in the module space. The lack of discovery, therefore, means they will not be run.